Revert "Feature/bc 5847"
