--- src/main/java/faang/school/accountservice/dto/User.java
@@ -1,11 +1,0 @@
-package faang.school.accountservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class User {
-    private Long id;
-    private String username;
-}
