--- src/main/java/faang/school/accountservice/enums/AccountStatus.java
@@ -1,18 +1,0 @@
-package faang.school.accountservice.enums;
-
-public enum AccountStatus {
-    OPEN("open"),
-    FROZEN("frozen"),
-    BLOCKED("blocked"),
-    CLOSED("closed"),
-    ;
-    private final String status;
-
-    AccountStatus(String status) {
-        this.status = status;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-}
