--- src/main/java/faang/school/accountservice/enums/AccountType.java
@@ -1,19 +1,0 @@
-package faang.school.accountservice.enums;
-
-public enum AccountType {
-    DEBIT("debit"),
-    CURRENCY("currency"),
-    CREDIT("credit"),
-    DEPOSIT("deposit"),
-    SAVINGS("savings"),
-    ;
-    private final String type;
-
-    AccountType(String type) {
-        this.type = type;
-    }
-
-    public String getType() {
-        return type;
-    }
-}
