--- src/main/java/faang/school/accountservice/enums/OwnerType.java
@@ -1,6 +1,0 @@
-package faang.school.accountservice.enums;
-
-public enum OwnerType {
-    USER,
-    PROJECT
-}
