--- src/main/java/faang/school/accountservice/exception/DataValidationException.java
@@ -1,7 +1,0 @@
-package faang.school.accountservice.exception;
-
-public class DataValidationException extends RuntimeException {
-    public DataValidationException(String message) {
-        super(message);
-    }
-}
