--- src/main/java/faang/school/accountservice/exception/IdempotencyException.java
@@ -1,7 +1,0 @@
-package faang.school.accountservice.exception;
-
-public class IdempotencyException extends RuntimeException {
-    public IdempotencyException(String message) {
-        super(message);
-    }
-}
