--- src/main/java/faang/school/accountservice/exception/LockedRequestException.java
@@ -1,7 +1,0 @@
-package faang.school.accountservice.exception;
-
-public class LockedRequestException extends RuntimeException {
-    public LockedRequestException(String message) {
-        super(message);
-    }
-}
