--- src/main/java/faang/school/accountservice/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package faang.school.accountservice.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
