--- src/main/java/faang/school/accountservice/exception/RequestNotFoundException.java
@@ -1,9 +1,0 @@
-package faang.school.accountservice.exception;
-
-import jakarta.persistence.EntityNotFoundException;
-
-public class RequestNotFoundException extends EntityNotFoundException {
-    public RequestNotFoundException(String message) {
-        super(message);
-    }
-}
