--- src/main/java/faang/school/accountservice/model/Account.java
@@ -1,63 +1,0 @@
-package faang.school.accountservice.model;
-
-import faang.school.accountservice.enums.Currency;
-import faang.school.accountservice.enums.AccountStatus;
-import faang.school.accountservice.enums.AccountType;
-import faang.school.accountservice.enums.OwnerType;
-import jakarta.persistence.*;
-import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Data
-@Table(name = "account")
-public class Account {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(name = "number", nullable = false, length = 20)
-    private String number;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "owner_type", nullable = false)
-    private OwnerType ownerType;
-
-    @Column(name = "owner_id", nullable = false)
-    private long ownerId;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "account_type", nullable = false)
-    private AccountType accountType;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "currency", nullable = false, length = 3)
-    private Currency currency;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "status", nullable = false)
-    private AccountStatus status;
-
-    @CreationTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "updated_at", nullable = false)
-    private LocalDateTime updatedAt;
-
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "closed_at")
-    private LocalDateTime closedAt;
-
-    @Version
-    @Column(name = "version", nullable = false)
-    private int version;
-}
-
