--- src/main/java/faang/school/accountservice/model/request/Request.java
@@ -1,72 +1,0 @@
-package faang.school.accountservice.model.request;
-
-import faang.school.accountservice.enums.OwnerType;
-import faang.school.accountservice.util.MapToJsonConverter;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.Instant;
-import java.util.Map;
-import java.util.UUID;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "request")
-public class Request {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(name = "idempotency_key", nullable = false, unique = true)
-    private UUID idempotencyKey;
-
-    @Column(name = "owner_id", nullable = false)
-    private Long ownerId;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "owner_type", nullable = false)
-    private OwnerType ownerType;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "request_type", nullable = false)
-    private RequestType requestType;
-
-    @Column(name = "lock_value", nullable = false)
-    private String lockValue;
-
-    @Column(name = "is_open", nullable = false)
-    private boolean isOpen;
-
-    @Column(name = "input_data")
-    @Convert(converter = MapToJsonConverter.class)
-    private Map<String, Object> inputData;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "status", nullable = false)
-    private RequestStatus requestStatus;
-
-    @Column(name = "status_details")
-    private String statusDetails;
-
-    @CreationTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "created_at", nullable = false)
-    private Instant createdAt;
-
-    @UpdateTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "updated_at", nullable = false)
-    private Instant updatedAt;
-
-    @Version
-    private Long version;
-}
