--- src/main/java/faang/school/accountservice/model/request/RequestType.java
@@ -1,10 +1,0 @@
-package faang.school.accountservice.model.request;
-
-public enum RequestType {
-    CREATE_ACCOUNT,
-    TRANSFER_FUNDS,
-    CLOSE_ACCOUNT,
-    DEPOSIT_FUNDS,
-    WITHDRAW_FUNDS,
-    UPDATE_ACCOUNT
-}
