--- src/main/java/faang/school/accountservice/repository/AccountRepository.java
@@ -1,9 +1,0 @@
-package faang.school.accountservice.repository;
-
-import faang.school.accountservice.model.Account;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
-}
