--- src/main/java/faang/school/accountservice/util/MapToJsonConverter.java
@@ -1,38 +1,0 @@
-package faang.school.accountservice.util;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-import lombok.RequiredArgsConstructor;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Converter
-@RequiredArgsConstructor
-@SuppressWarnings("unchecked")
-public class MapToJsonConverter implements AttributeConverter<Map<String, Object>, String> {
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public String convertToDatabaseColumn(Map<String, Object> attribute) {
-        try {
-            return objectMapper.writeValueAsString(attribute);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Override
-    public Map<String, Object> convertToEntityAttribute(String dbData) {
-        if (dbData == null) {
-            return new HashMap<>();
-        }
-        try {
-            return objectMapper.readValue(dbData, Map.class);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
