+++ src/main/resources/application.yaml
@@ -23,17 +23,3 @@ server:
 logging:
   level:
     root: info
-
-feign:
-  client:
-    config:
-      default:
-        connectTimeout: 5000
-        readTimeout: 5000
-
-user-service:
-  host: localhost
-  port: 8080
-project-service:
-  host: localhost
-  port: 8082
