+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,7 +1,3 @@
 databaseChangeLog:
   - include:
       file: db/changelog/changeset/V001__account-service_account.sql
-  - include:
-      file: db/changelog/changeset/V002__account-service_payment-account.sql
-  - include:
-      file: db/changelog/changeset/V003__request.sql
