+++ src/test/java/faang/school/accountservice/AccountServiceApplicationTests.java
@@ -3,9 +3,6 @@ package faang.school.accountservice;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.List;
-
 
 class AccountServiceApplicationTests {
     @Test
